Changes to Logic, changed data-types for sensor values

+++ src/discoverylab/telebot/master/arms/mapper/Mapper.java
@@ -2,7 +2,7 @@ package discoverylab.telebot.master.arms.mapper;
 
 public abstract class Mapper {
 	
-	public abstract long constrain(long val, long servo_max, long servo_min);
+	public abstract int constrain(int val, int servo_max, int servo_min);
 	
-	public abstract long map(long val, long sensor_max, long sensor_min, long servo_max, long servo_min);
+	public abstract int map(int val, int sensor_max, int sensor_min, int servo_max, int servo_min);
 }
