+++ src/discoverylab/telebot/master/arms/mapper/ServoDataMapper.java
@@ -7,9 +7,9 @@ import discoverylab.telebot.master.arms.model.YEIDataModel;
 public class ServoDataMapper extends Mapper{
 	public static String TAG = makeLogTag(ServoDataMapper.class);
 	@Override
-	public long constrain(long val, long servo_max, long servo_min)
+	public int constrain(int val, int servo_max, int servo_min)
 	{
-		long ret = val;
+		int ret = val;
 
 		if(val <= servo_min)
 		{
@@ -24,11 +24,11 @@ public class ServoDataMapper extends Mapper{
 	}
 	
 	@Override
-	public long map(long val, long sensor_max, long sensor_min, long servo_max, long servo_min)
+	public int map(int val, int sensor_max, int sensor_min, int servo_max, int servo_min)
 	{
-        long mapped = (val - sensor_min) * (servo_max - servo_min) / (sensor_max - sensor_min) + servo_min;
+        int mapped = (val - sensor_min) * (servo_max - servo_min) / (sensor_max - sensor_min) + servo_min;
         
-        long constrained =  constrain(mapped, servo_max, servo_min);
+        int constrained =  constrain(mapped, servo_max, servo_min);
         
         return constrained;
 	}
