+++ src/discoverylab/telebot/master/arms/model/YEIDataModel.java
@@ -5,9 +5,9 @@ public class YEIDataModel {
 	public YEIDataModel(){	
 	}
 	
-	public YEIDataModel(String jointType, long x, long maxX, long minX, 
-			long y, long maxY, long minY, 
-			long z, long maxZ, long minZ){
+	public YEIDataModel(String jointType, int x, int maxX, int minX, 
+			int y, int maxY, int minY, 
+			int z, int maxZ, int minZ){
 		this.jointType = jointType;
 		this.x = x;
 		this.maxX = maxX;
@@ -26,70 +26,70 @@ public class YEIDataModel {
 	public void setJointType(String jointType) {
 		this.jointType = jointType;
 	}
-	public long getX() {
+	public int getX() {
 		return x;
 	}
-	public void setX(long x) {
+	public void setX(int x) {
 		this.x = x;
 	}
-	public long getMaxX() {
+	public int getMaxX() {
 		return maxX;
 	}
-	public void setMaxX(long x) {
+	public void setMaxX(int x) {
 		this.maxX = x;
 	}
-	public long getMinX() {
+	public int getMinX() {
 		return minX;
 	}
-	public void setMinX(long x) {
+	public void setMinX(int x) {
 		this.minX = x;
 	}
-	public long getY() {
+	public int getY() {
 		return y;
 	}
-	public void setY(long y) {
+	public void setY(int y) {
 		this.y = y;
 	}
-	public long getMaxY() {
+	public int getMaxY() {
 		return maxY;
 	}
-	public void setMaxY(long x) {
+	public void setMaxY(int x) {
 		this.maxY = x;
 	}
-	public long getMinY() {
+	public int getMinY() {
 		return minY;
 	}
-	public void setMinY(long x) {
+	public void setMinY(int x) {
 		this.minY = x;
 	}
-	public long getZ() {
+	public int getZ() {
 		return z;
 	}
-	public void setZ(long z) {
+	public void setZ(int z) {
 		this.z = z;
 	}
-	public long getMaxZ() {
+	public int getMaxZ() {
 		return maxZ;
 	}
-	public void setMaxZ(long x) {
+	public void setMaxZ(int x) {
 		this.maxZ = x;
 	}
-	public long getMinZ() {
+	public int getMinZ() {
 		return minZ;
 	}
-	public void setMinZ(long x) {
+	public void setMinZ(int x) {
 		this.minZ = x;
 	}
 	private String jointType;
-	private long x;
-	private long maxX;
-	private long minX;
-	private long y;
-	private long maxY;
-	private long minY;
-	private long z;
-	private long maxZ;
-	private long minZ;
+	private int x;
+	private int maxX;
+	private int minX;
+	private int y;
+	private int maxY;
+	private int minY;
+	private int z;
+	private int maxZ;
+	private int minZ;
 	
 
 }
